+++ src/handler/filter.rs
@@ -83,6 +83,7 @@ where
             obligations: Pred::obligations(),
             guaranteed_outcomes: BTreeSet::default(),
             conditional_outcomes: BTreeSet::default(),
+            continues: true,
         },
     )
 }