+++ src/handler/filter_map.rs
@@ -96,6 +96,7 @@ where
             obligations: Projection::obligations(),
             guaranteed_outcomes: BTreeSet::from_iter(vec![Type::of::<NewType>()]),
             conditional_outcomes: BTreeSet::default(),
+            continues: true,
         },
     )
 }